Add basic tabbed workflow structure
