+++ src/App.css
@@ -1093,3 +1093,21 @@ body.dark-mode .spec-input-column label {
     grid-template-columns: 1fr;
   }
 }
+/* Navigation tabs */
+.main-nav {
+  display: flex;
+  gap: 8px;
+  padding: 10px;
+  background: #f2f2f2;
+  border-bottom: 1px solid #ccc;
+}
+.main-nav button {
+  background: none;
+  border: 1px solid #ccc;
+  padding: 6px 12px;
+  cursor: pointer;
+}
+.main-nav button:hover {
+  background: #e6e6e6;
+}
+
