+++ src/pages/AnalysisInfoPage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function AnalysisInfoPage() {
+  return (
+    <div className="content-area">
+      <h2>Analysis Info</h2>
+      <p>This section will capture equipment and analyst information.</p>
+    </div>
+  );
+}
