+++ src/pages/NotesPage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function NotesPage() {
+  return (
+    <div className="content-area">
+      <h2>Notes</h2>
+      <p>Use this section to record analysis notes.</p>
+    </div>
+  );
+}
