+++ src/pages/RiskAssumedReopPage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function RiskAssumedReopPage() {
+  return (
+    <div className="content-area">
+      <h2>Risk With Assumed REOP</h2>
+      <p>Risk analysis workflows will be implemented here.</p>
+    </div>
+  );
+}
